bugfix - pub/sub and simultaniously checking server vals
threw an exception when the subscription handler routine was running after a datachange and when the fhs connection thread was running in its while loop, getting values from the same server. so a second connection was created, one for the subhandler and one for the checking if the server is running inside the while loop in the fhs thread.